pyproject.toml
[tool.poetry]
package-mode = false
name = "living-wallpaper"
version = "0.0.0"
description = "Change lil' things in your wallpaper"
authors = ["Štěpán Žák <stepanzak@proton.me>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyora = "^0.3.11"
pillow = "^10.3.0"
tomli = "^2.0.1"
argparse = "^1.4.0"
pathlib = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
